Extract Tag from Users Table
